set pct to int

+++ src/main/java/org/matsim/prepare/longDistanceFreightGER/GenerateFreightPlans.java
@@ -59,7 +59,7 @@ public class GenerateFreightPlans implements MATSimAppCommand {
     private int workingDays;
 
     @CommandLine.Option(names = "--sample", defaultValue = "100", description = "Sample size of the freight plans (0, 100]")
-    private double pct;
+    private int pct;
 
 	@CommandLine.Option(names = "--land-use-filter", description = "specify land use type to filter out starting locations. Empty means no filter",
 		arity = "0..*", split = ",", defaultValue = "industrial,commercial,retail")
@@ -88,7 +88,7 @@ public class GenerateFreightPlans implements MATSimAppCommand {
 		if (!landUseTypes.isEmpty()){
 			landuse = new LanduseOptions(output.toString() + "/landuse-shp/landuse.shp", landUseTypes);
 		}
-        FreightAgentGenerator freightAgentGenerator = new FreightAgentGenerator(network, shpPath, landuse, averageTruckLoad, workingDays, pct / 100);
+        FreightAgentGenerator freightAgentGenerator = new FreightAgentGenerator(network, shpPath, landuse, averageTruckLoad, workingDays, (double) pct / 100);
         log.info("Freight agent generator successfully created!");
 
         log.info("Reading trip relations...");
